Cirrus: Fix version-check to only run on `main` job
A conditional in `version-check` bypasses the test for PRs.  However,
it appears it was intended to execute during the daily cirrus-cron runs.
However, the cron-job it references (`nightly`) doesn't exist.  This is
causing the test to run for every merge into `main`, and never run for
`main` branch cirrus-cron job.  Fix the name so the test **ONLY**
runs for the `main` branch cron-job.

Also, since the test is currently failing, update the docs as per the
output instructions.

Signed-off-by: Chris Evich <cevich@redhat.com>