[v4.4.1-rhel] [CI:DOCS] Touch up secret man page
This is a working theory fix.  The man pages including the
secret.md option are not resolving properly.  The secret.md
file is the only one with a text line with three back ticks on the start
and end of the line.  Elsewhere we have the backticks on separate lines
and the text in it's own line.

This might not fix the issue, but at the very least it makes things
consistent.

[NO NEW TESTS NEEDED]

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>