[CI:DOCS] Add image not found info to troubleshooting
Addresses: https://bugzilla.redhat.com/show_bug.cgi?id=2139191

We have a timing issue when doing a build or pull of an image and
a second process is removing them at the same time.  This is not
a window we're going to close now, so I'm adding a note to the
troubleshooting guide to document it.

[NO NEW TESTS NEEDED]
Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>