test/e2e: use fresh network config dir for each test
The e2e test are isolated and have their own --root/--runroot arguments.
However networks were always shared, this causes problem with tests that
do a prune or reset because they can effect other parallel running
tests.

Over the time I fixed some of of these cases to use their own config dir
but #17946 suggests that this is not enough. Instead of trying to find
and fix these tests just go with the big hammer and make every test use
a new clean network config directory.

This will also make the use of `defer podmanTest.removeNetwork(...)`
unnecessary. This is required at the moment for every test which creates
a network. However to keep the diff small and to see if it is even
working I will do it later in a follow up commit.

Fixes #17946

Signed-off-by: Paul Holzinger <pholzing@redhat.com>